ai-agent-pf/model/requirements.txt
azure-ai-projects==1.0.0b8
azure-identity 